Make post default order optional positive integer
Posts on the same page should have consecutive default order values,
with 1 for the bottommost post and *n* for the topmost post, where *n*
is the number of posts on the page.

A new or edited post is allowed to temporarily have
`:post/default-order` be `nil` or omitted, if the post is meant to
be displayed last on the page. Once the post is sent, however, it and
other posts on the same page should be re-sorted to have consecutive
default order values.