Add custom post order for each page (#177)
* Add field for default order of posts to backend

This commit introduces a new field, `:post/default-order`, that
determines how posts should be ordered when no sorting methods are set.
Posts with higher `:post/default-order` should be positioned higher up,
closer to the top of the page.

* Make post default order optional positive integer

Posts on the same page should have consecutive default order values,
with 1 for the bottommost post and *n* for the topmost post, where *n*
is the number of posts on the page.

A new or edited post is allowed to temporarily have
`:post/default-order` be `nil` or omitted, if the post is meant to
be displayed last on the page. Once the post is sent, however, it and
other posts on the same page should be re-sorted to have consecutive
default order values.

* Sort posts on post submissions and removals

FIXME: 15/65 frontend test items are failing.

* Fix backend tests

* Make default post order values start at 0

This commit makes 0 the value for the bottommost post instead of 1,
which might help mitigate off-by-one errors.

* Add property-based tests for post sorting

They run. That's good. They also add 100 lines of code and 300 test
assertions.

* Remove property-based tests for post sorting

* Change post sorting method

A new or edited post with `:post/default-order n` is now guaranteed to
be inserted at the *n*-th position (zero-based). Posts with
`:post/default-order` being `nil` or too high are added at the end.

* Fix post sorting; add tests

* Remove unused sample data and tests

* Simplify post sorting

This commit simplifies post sorting using lazy sequence operations, as
opposed to incrementing/decrementing indices directly.

* Undo whitespace changes