Make front end display pages' custom post orders (#186)
* Add default post order to query pattern

This commit allows the front end to pull and display pages' default post
orders correctly.

It does not yet allow new posts to have custom positions. New posts are
still automatically added to the end by default.

* Remove use of as-> macro

This commit removes the use of the `as->` macro in favor of explicit
`let` bindings.